refactor Word location
